pyproject.toml
[project]
name = "sipgiris"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.2.1",
]
